move send email components into event component
