+++ src/views/EventView/index.tsx
@@ -1,11 +1,10 @@
 'use client';
 import Event from '@/components/Event';
-import EmailEditor from '@/components/EmailEditor';
 import { EventResponse } from '@/types/dataModel/event';
 import { EventVolunteerResponse } from '@/types/dataModel/eventVolunteer';
 import { Box, Button } from '@mui/material';
 import Link from 'next/link';
-import React, { useState } from 'react';
+import React from 'react';
 
 export default function EventView({
   event,
@@ -14,11 +13,6 @@ export default function EventView({
   event: EventResponse;
   eventVolunteers: EventVolunteerResponse[];
 }) {
-  const [showEmail, setShowEmail] = useState(false);
-
-  const handleClick = () => {
-    setShowEmail(true);
-  };
   return (
     <Box>
       {/* TODO: Add back in after demo/implementation */}
@@ -28,11 +22,8 @@ export default function EventView({
           Check-in
         </Button>
       </Link>
-      <Button variant="contained" fullWidth onClick={handleClick}>
-        New Email
-      </Button>
+
       <Event event={event} eventVolunteers={eventVolunteers} />
-      {showEmail && <EmailEditor event={event} volunteers={eventVolunteers} />}
     </Box>
   );
 }
